+++ app/oauth2-server-config/client/admin/startup.js
@@ -1,5 +1,5 @@
 import { hasAllPermission } from '../../../authorization';
-import { registerAdminSidebarItem } from '../../../ui-admin/client';
+import { registerAdminSidebarItem } from '../../../../client/admin';
 
 registerAdminSidebarItem({
 	href: 'admin-oauth-apps',
