+++ client/importPackages.js
@@ -72,7 +72,6 @@ import '../app/spotify/client';
 import '../app/tokenpass/client';
 import '../app/ui';
 import '../app/ui-account';
-import '../app/ui-admin/client';
 import '../app/ui-clean-history';
 import '../app/ui-flextab';
 import '../app/ui-login';
