+++ client/main.js
@@ -28,3 +28,4 @@ import './startup/startup';
 import './startup/unread';
 import './startup/userSetUtcOffset';
 import './startup/usersObserve';
+import './admin';
